# api-docs.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "Monitoring Platform API",
    "description": "API for system monitoring, alerts, notifications, and reporting",
    "contact": {
      "name": "API Support",
      "email": "support@example.com"
    },
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://localhost:8081",
      "description": "Development"
    }
  ],
...
  "paths": {
...
    "/api/notifications/settings": {
      "get": {
        "tags": [
          "Notifications"
        ],
        "summary": "Get notification settings",
        "description": "Get user notification preferences",
        "operationId": "getSettings",
        "parameters": [
          {
            "name": "userId",
            "in": "query",
            "description": "User ID",
            "required": false,
            "schema": {
              "type": "string",
              "default": "user123"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "type": "object",
                  "additionalProperties": {
                    "type": "object"
                  }
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "Notifications"
        ],
        "summary": "Update notification settings",
        "description": "Update user notification preferences",
        "operationId": "updateSettings",
        "parameters": [
          {
            "name": "userId",
            "in": "query",
            "description": "User ID",
            "required": false,
            "schema": {
              "type": "string",
              "default": "user123"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "additionalProperties": {
                  "type": "object"
                }
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "type": "object",
                  "additionalProperties": {
                    "type": "object"
                  }
                }
              }
            }
          }
        }
      }
    },
...
    "/api/notifications/webhook/secureProcess": {
      "post": {
        "tags": [
          "Notifications"
        ],
        "summary": "Process webhook (secure)",
        "description": "Process XML webhook with XXE protection",
        "operationId": "secureProcessWebhook",
        "requestBody": {
          "content": {
            "application/xml": {
              "schema": {
                "type": "string"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "type": "object",
                  "additionalProperties": {
                    "type": "object"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/notifications/webhook/process": {
      "post": {
        "tags": [
          "Notifications"
        ],
        "summary": "Process webhook",
        "description": "Process incoming XML webhook notification",
        "operationId": "processWebhook",
        "requestBody": {
          "content": {
            "application/xml": {
              "schema": {
                "type": "string"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "type": "object",
                  "additionalProperties": {
                    "type": "object"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/notifications/test": {
      "post": {
        "tags": [
          "Notifications"
        ],
        "summary": "Send test notification",
        "description": "Send a test notification",
        "operationId": "sendTest",
        "parameters": [
          {
            "name": "userId",
            "in": "query",
            "description": "User ID",
            "required": false,
            "schema": {
              "type": "string",
              "default": "user123"
            }
          },
          {
            "name": "channel",
            "in": "query",
            "description": "Channel",
            "required": false,
            "schema": {
              "type": "string",
              "default": "email"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "type": "object",
                  "additionalProperties": {
                    "type": "object"
                  }
                }
              }
            }
          }
        }
      }
    },
...
    "/api/reports/secure-network-check": {
      "get": {
        "tags": [
          "Reports"
        ],
        "summary": "Network check (secure)",
        "description": "Run ping safely without shell",
        "operationId": "secureNetworkCheck",
        "parameters": [
          {
            "name": "hostname",
            "in": "query",
            "description": "Hostname to check",
            "required": false,
            "schema": {
              "type": "string",
              "default": "localhost"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "type": "object",
                  "additionalProperties": {
                    "type": "object"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/reports/network-check": {
      "get": {
        "tags": [
          "Reports"
        ],
        "summary": "Network connectivity check",
        "description": "Run ping to check host connectivity",
        "operationId": "networkCheck",
        "parameters": [
          {
            "name": "hostname",
            "in": "query",
            "description": "Hostname to check",
            "required": false,
            "schema": {
              "type": "string",
              "default": "localhost"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "type": "object",
                  "additionalProperties": {
                    "type": "object"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/notifications/history": {
      "get": {
        "tags": [
          "Notifications"
        ],
        "summary": "Get notification history",
        "description": "Get recent notifications for user",
        "operationId": "getHistory",
        "parameters": [
          {
            "name": "userId",
            "in": "query",
            "description": "User ID",
            "required": false,
            "schema": {
              "type": "string",
              "default": "user123"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "description": "Limit",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 10
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "type": "object",
                  "additionalProperties": {
                    "type": "object"
                  }
                }
              }
            }
          }
        }
      }
    },
...
  },
...
}